Add rm_techreq and rm_default_techreq
rm_techreq is a briefing leader console command which excepts 0 or 1,
1 is default behavior. If 0 all marines can hack doors and computers,
tech isn't needed. Mission doesn't restart if tech dies

rm_default_techreq is a config cvar which is used to set the default
value of rm_techreq. It is only in effect when rm_default_game_mode 6
(Server Default Game Mode)